Design QueryClient.DoTx method
